requirements.txt
asgiref==3.8.1
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-ses==3.5.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docker==7.0.0
idna==3.6
jmespath==1.0.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
